refactor(网络): 将消息头长度属性类型从int改为ushort和uint
统一消息头长度属性的数据类型，提高类型安全性并明确数值范围。在DefaultMessageDecoderHandler中添加缺失的PackageHeaderLength属性实现。

+++ GameFrameX.NetWork.Abstractions/IMessageDecoderHandler.cs
@@ -10,7 +10,7 @@ public interface IMessageDecoderHandler
     /// <summary>
     /// 消息头长度
     /// </summary>
-    int PackageHeaderLength { get; }
+    ushort PackageHeaderLength { get; }
 
     /// <summary>
     /// 解析消息
